Add junit-jupiter-params to POM test scope
Per discussion with Dave and re-reviewing the tutorial section on steps
for installation I had missed, add that dependency to the POM for #178.

Also, per his recommendation, add it to the test scope and do not set an
explicit version number because the core JUnit5 version is handled by the
support POM for oss-maven project shared across the different Java
projects.